Split unit tests by method types
